Use alternatives to specify the default version
We use the alternatives system to allow system admins to specify the
default drbd-utils version on the system. While upstream has implemented
a fallback mechanism, this has some rough edges and may cause issues
with scripts e.g. trying DRBD 8.3 commands.

We provide three alternatives and default to 8.4 being the default
utilities and manpage version, as this is the DRBD kernel module version
currently shipped with Debian.